SCP-1921: ChainIndex: tidying up
- add ChainIndex module
- move some declarations to Type module
- add ChainIndexUrl newtype wrapper for BaseUrl in config
- explicit exports in type module
